+++ release-notes.md
@@ -6,6 +6,7 @@ For players:
 * The SMAPI console is now much simpler and easier to read.
 * The SMAPI console now adjusts its colors when you have a light terminal background.
 * Updated compatibility list.
+* Improved errors when a mod DLL can't be loaded.
 
 For mod developers:
 * Added new APIs to edit, inject, and reload XNB assets loaded by the game at any time.